Add method in initialize_quizes
